requirements.txt
google-cloud-bigquery==3.13.0
pandas==2.1.4
